playground/pyproject.toml
[project]
name = "sangria-net"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cdp-sdk>=1.39.1",
    "fastapi-x402>=0.1.8",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.41.0",
    "x402[httpx]>=2.3.0",
]
